Change IFastEnumerator so that it enumerates over the DATA rather than the KEY (implements IEnumerator<TData> instead of IEnumerator<TKey>)
An IFastEnumerator is an enumerator that can skip ahead in the data structure more quickly, using a key and the MoveUntilGreaterThanOrEqual() method.

So it skips ahead by KEYS. But the thing it actually returns, it makes more sense to think of it as enumerating over the DATA. The data is the actual thing you want. The keys are how you traverse the structure, but the data is what you are after, so therefore IEnumerator<TData> makes more sense intuitively. The logic of the enumerator doesn't change. It's just a question of what type do you want the "Current" property to return.